pyproject.toml
[tool.poetry]
name = "neuralprophet"
version = "1.0.0rc10"
description = "NeuralProphet is an easy to learn framework for interpretable time series forecasting."
authors = ["Oskar Triebe <triebe@stanford.edu>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
]

[tool.poetry.urls]
Homepage = "https://github.com/ourownstory/neural_prophet"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = ">=1.25.0,<2.0.0"
pandas = ">=2.0.0"
torch = ">=2.0.0,<2.4.0"
# Note: torch defaults to already installed version or installs CUDA version
# If you want CPU-only torch, install that before installing neuralprophet.
pytorch-lightning = ">=2.0.0,<2.4.0"
tensorboard = ">=2.11.2"
torchmetrics = ">=1.0.0"
typing-extensions = ">=4.5.0"
holidays = ">=0.41,<1.0"
captum = ">=0.6.0"
matplotlib = ">=3.5.3"
plotly = ">=5.13.1"
kaleido = "0.2.1" # required for plotly static image export
plotly-resampler = { version = ">=0.9.2", optional = true }
livelossplot = { version = ">=0.5.5", optional = true }
lightning-fabric = ">=2.0.0,<2.4.0"
pyyaml = ">=6.0"
pydantic = ">=1.10,<3.0"
sqlalchemy = ">=2.0,<3.0"
fastapi = ">=0.110,<0.112"
uvicorn = ">=0.23,<0.30"
streamlit = ">=1.32,<1.40"
jinja2 = ">=3.1,<4.0"
redis = ">=5.0,<6.0"
apscheduler = ">=3.10,<4.0"
prometheus-client = ">=0.19,<1.0"

[tool.poetry.extras]
plotly-resampler = ["plotly-resampler"]
live = ["livelossplot"]

[tool.poetry.group.dev.dependencies] # For dev involving notebooks
ipykernel = ">=6.29.2"
nbformat = ">=5.8.0"

[tool.poetry.group.pytest] # pytest dev setup and CI
optional = true

[tool.poetry.group.pytest.dependencies] 
pytest = "^8.0.0"
pytest-xdist = "^3.1.0"
pytest-cov = "^4.1.0"

[tool.poetry.group.docs] # for building docs; and testing docs CI
optional = true

[tool.poetry.group.docs.dependencies]
myst-parser = "^2.0.0"
nbsphinx = "^0.9.0"
nbsphinx-link = "^1.3.0"
sphinx = "^7.0.0"
sphinx-fontawesome = "^0.0.6"
furo = "^2024.1.29"

[tool.poetry.group.metrics] # for metrics CI
optional = true

[tool.poetry.group.metrics.dependencies]
tabulate = "^0.9.0" # Used in model metrics CI only; md export for github-actions bot

[tool.poetry.group.linters] # for linters CI
optional = true

[tool.poetry.group.linters.dependencies]
black = { extras = ["jupyter"], version = "^24.1.0" }
isort = "^5.12.0"
pandas-stubs = "^2.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | example_data
    | example_notebooks
    | notes
    | site
  )/
  | .gitignore
)
'''

[tool.isort]
profile = "black"
line_length = 120

[tool.pyright]
include = [
    "neuralprophet/**/*.py"
]
exclude = [
    "neuralprophet/**/*plot*.py"
]

[tool.ruff]
line-length = 120
typing-modules = ["neuralprophet.np_types"]
